move spi functions to main
This disables the modules which were using spi functions directly or
indirectly.
